separate pidfile package from main
