+++ main.go
@@ -3,13 +3,10 @@ package main
 import (
 	"flag"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"os/signal"
-	"path/filepath"
 	"regexp"
 	"runtime"
-	"strconv"
 	"strings"
 	"syscall"
 	"time"
@@ -17,7 +14,7 @@ import (
 	"github.com/mackerelio/mackerel-agent/command"
 	"github.com/mackerelio/mackerel-agent/config"
 	"github.com/mackerelio/mackerel-agent/logging"
-	"github.com/mackerelio/mackerel-agent/util"
+	"github.com/mackerelio/mackerel-agent/pidfile"
 	"github.com/mackerelio/mackerel-agent/version"
 	"github.com/motemen/go-cli"
 )
@@ -160,45 +157,6 @@ func resolveConfig(fs *flag.FlagSet, argv []string) (*config.Config, error) {
 	return conf, nil
 }
 
-func createPidFile(pidfile string) error {
-	if pidfile == "" {
-		return nil
-	}
-	if pidString, err := ioutil.ReadFile(pidfile); err == nil {
-		if pid, err := strconv.Atoi(string(pidString)); err == nil {
-			if util.ExistsPid(pid) {
-				return fmt.Errorf("pidfile found, try stopping another running mackerel-agent or delete %s", pidfile)
-			}
-			// Note mackerel-agent in windows can't remove pidfile during stoping the service
-			logger.Warningf("Pidfile found, but there seems no another process of mackerel-agent. Ignoring %s", pidfile)
-		} else {
-			logger.Warningf("Malformed pidfile found. Ignoring %s", pidfile)
-		}
-	}
-
-	err := os.MkdirAll(filepath.Dir(pidfile), 0755)
-	if err != nil {
-		return err
-	}
-	file, err := os.Create(pidfile)
-	if err != nil {
-		return err
-	}
-	defer file.Close()
-
-	_, err = fmt.Fprintf(file, "%d", os.Getpid())
-	return err
-}
-
-func removePidFile(pidfile string) {
-	if pidfile == "" {
-		return
-	}
-	if err := os.Remove(pidfile); err != nil {
-		logger.Errorf("Failed to remove the pidfile: %s: %s", pidfile, err)
-	}
-}
-
 func setLogLevel(silent, verbose bool) {
 	if silent {
 		logging.SetLogLevel(logging.ERROR)
@@ -212,10 +170,10 @@ func start(conf *config.Config, termCh chan struct{}) error {
 	setLogLevel(conf.Silent, conf.Verbose)
 	logger.Infof("Starting mackerel-agent version:%s, rev:%s, apibase:%s", version.VERSION, version.GITCOMMIT, conf.Apibase)
 
-	if err := createPidFile(conf.Pidfile); err != nil {
-		return fmt.Errorf("createPidFile(%q) failed: %s", conf.Pidfile, err)
+	if err := pidfile.Create(conf.Pidfile); err != nil {
+		return fmt.Errorf("pidfile.Create(%q) failed: %s", conf.Pidfile, err)
 	}
-	defer removePidFile(conf.Pidfile)
+	defer pidfile.Remove(conf.Pidfile)
 
 	ctx, err := command.Prepare(conf)
 	if err != nil {
