+++ main_test.go
@@ -14,6 +14,7 @@ import (
 	"time"
 
 	"github.com/mackerelio/mackerel-agent/command"
+	"github.com/mackerelio/mackerel-agent/pidfile"
 )
 
 func TestParseFlags(t *testing.T) {
@@ -130,25 +131,25 @@ func TestCreateAndRemovePidFile(t *testing.T) {
 	fpath := file.Name()
 	defer os.Remove(fpath)
 
-	err = createPidFile(fpath)
+	err = pidfile.Create(fpath)
 	if err != nil {
 		t.Errorf("pid file should be created but, %s", err)
 	}
 
 	if runtime.GOOS != "windows" {
-		if err := createPidFile(fpath); err == nil || !strings.HasPrefix(err.Error(), "pidfile found, try stopping another running mackerel-agent or delete") {
+		if err := pidfile.Create(fpath); err == nil || !strings.HasPrefix(err.Error(), "pidfile found, try stopping another running mackerel-agent or delete") {
 			t.Errorf("creating pid file should be failed when the running process exists, %s", err)
 		}
 	}
 
-	removePidFile(fpath)
-	if err := createPidFile(fpath); err != nil {
+	pidfile.Remove(fpath)
+	if err := pidfile.Create(fpath); err != nil {
 		t.Errorf("pid file should be created but, %s", err)
 	}
 
-	removePidFile(fpath)
+	pidfile.Remove(fpath)
 	ioutil.WriteFile(fpath, []byte(fmt.Sprint(math.MaxInt32)), 0644)
-	if err := createPidFile(fpath); err != nil {
+	if err := pidfile.Create(fpath); err != nil {
 		t.Errorf("old pid file should be ignored and new pid file should be created but, %s", err)
 	}
 }
