+++ pidfile/pid_darwin.go
@@ -1,4 +1,4 @@
-package util
+package pidfile
 
 import (
 	"fmt"
