+++ pidfile/pid_windows.go
@@ -1,4 +1,4 @@
-package util
+package pidfile
 
 func existsPid(_ int) bool {
 	// XXX not implemented. should use `tasklist` command or so
